+++ apps/src/templates/teacherNavigation/UnitCalendar.tsx
@@ -6,7 +6,10 @@ import UnitCalendarGrid from '@cdo/apps//code-studio/components/progress/UnitCal
 import {initializeRedux} from '@cdo/apps/code-studio/components/progress/TeacherUnitOverview';
 import {getAuthenticityToken} from '@cdo/apps/util/AuthenticityTokenStore';
 import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
+import {SimpleDropdown} from '@cdo/apps/componentLibrary/dropdown';
 import i18n from '@cdo/locale';
+
+import styles from './teacher-navigation.module.scss';
 
 const WEEKLY_INSTRUCTIONAL_MINUTES_OPTIONS = [
   45, 90, 135, 180, 225, 270, 315, 360, 405, 450,
@@ -17,7 +20,7 @@ const UnitCalendar: React.FC = () => {
   const [isLoading, setIsLoading] = useState(false); // it is only loading when you do the fetch
 
   const [weeklyInstructionalMinutes, setWeeklyInstructionalMinutes] =
-    useState<number>(WEEKLY_INSTRUCTIONAL_MINUTES_OPTIONS[4]);
+    useState<string>(WEEKLY_INSTRUCTIONAL_MINUTES_OPTIONS[4].toString());
 
   const unitName = useSelector(
     (state: {unitSelection: {unitName: string}}) => state.unitSelection.unitName
@@ -78,34 +81,33 @@ const UnitCalendar: React.FC = () => {
     dispatch,
   ]);
 
-  const generateDropdownOptions = () => {
-    const options = WEEKLY_INSTRUCTIONAL_MINUTES_OPTIONS;
-    if (!options.includes(weeklyInstructionalMinutes)) {
-      options.push(weeklyInstructionalMinutes);
-    }
-    options.sort((a, b) => a - b);
-    return options.map(val => (
-      <option value={parseInt(val.toString())} key={`minutes-${val}`}>
-        {i18n.minutesLabel({number: val})}
-      </option>
-    ));
+  const weeklyMinutesOptions = WEEKLY_INSTRUCTIONAL_MINUTES_OPTIONS.map(
+    value => ({
+      value: value.toString(),
+      text: i18n.minutesLabel({number: value}),
+    })
+  );
+
+  const handleDropdownChange = (value: string) => {
+    setWeeklyInstructionalMinutes(value);
   };
 
   return (
-    <div style={styles.contentContainer}>
-      <div style={styles.minutesPerWeekWrapper}>
-        <div style={styles.minutesPerWeekDescription}>
+    <div className={styles.calendarContentContainer}>
+      <div className={styles.minutesPerWeekWrapper}>
+        <div className={styles.minutesPerWeekDescription}>
           {i18n.instructionalMinutesPerWeek()}
         </div>
-        <select
-          onChange={e =>
-            setWeeklyInstructionalMinutes(parseInt(e.target.value))
-          }
-          value={weeklyInstructionalMinutes}
-          style={styles.dropdown}
-        >
-          {generateDropdownOptions()}
-        </select>
+        <SimpleDropdown
+          name="minutesPerWeek"
+          onChange={event => handleDropdownChange(event.target.value)}
+          items={weeklyMinutesOptions}
+          selectedValue={weeklyInstructionalMinutes}
+          size="s"
+          dropdownTextThickness="thin"
+          labelText="minutes per week dropdown"
+          isLabelVisible={false}
+        />
       </div>
       {!isLoading && hasCalendar && (
         <UnitCalendarGrid
@@ -119,32 +121,3 @@ const UnitCalendar: React.FC = () => {
 };
 
 export default UnitCalendar;
-
-const styles = {
-  dialog: {
-    textAlign: 'left',
-    paddingLeft: 20,
-    paddingRight: 20,
-    paddingBottom: 20,
-  },
-  button: {
-    float: 'right',
-    marginTop: 30,
-  },
-  dropdown: {
-    width: 'fit-content',
-    marginBottom: 0,
-  },
-  minutesPerWeekWrapper: {
-    display: 'flex',
-    marginBottom: 10,
-    alignItems: 'center',
-  },
-  minutesPerWeekDescription: {
-    fontWeight: 'bold' as const,
-    marginRight: 10,
-  },
-  contentContainer: {
-    width: 'fit-content',
-  },
-};