+++ apps/test/unit/codebridge/FileBrowser/prompts/openNewFolderPromptTest.ts
@@ -1,30 +1,12 @@
 import {NewFolderFunction} from '@codebridge/codebridgeContext/types';
 import {DEFAULT_FOLDER_ID} from '@codebridge/constants';
-import {
-  validateNewFolderName,
-  openNewFolderPrompt,
-} from '@codebridge/FileBrowser/prompts/openNewFolderPrompt';
+import {openNewFolderPrompt} from '@codebridge/FileBrowser/prompts/openNewFolderPrompt';
 import {FolderId, ProjectFolder} from '@codebridge/types';
 
-import codebridgeI18n from '@cdo/apps/codebridge/locale';
-import {DialogControlInterface} from '@cdo/apps/lab2/views/dialogs';
-import {GenericPromptProps} from '@cdo/apps/lab2/views/dialogs/GenericPrompt';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 
-import testProject from '../../testProject.json';
-
-const getDialogControlMock = (
-  newFolderName: string
-): Pick<DialogControlInterface, 'showDialog'> => ({
-  showDialog: ({validateInput}: GenericPromptProps) => {
-    const error = validateInput?.(newFolderName);
-    if (error) {
-      return Promise.resolve({type: 'cancel', args: error});
-    } else {
-      return Promise.resolve({type: 'confirm', args: newFolderName});
-    }
-  },
-});
+import {testProject} from '../../test-files/';
+import {getDialogControlMock, getAnalyticsMock} from '../../test_utils';
 
 const getNewFolderMock = (
   parentId: FolderId
@@ -39,47 +21,10 @@ const getNewFolderMock = (
   return [newFolderData, mock];
 };
 
-type AnalyticsDataType = {event: string};
-type AnalyticsMockType = (event: string) => void;
-
-const getAnalyticsMock = (): [AnalyticsDataType, AnalyticsMockType] => {
-  const analyticsData = {} as AnalyticsDataType;
-  const mock = (event: string) => {
-    analyticsData.event = event;
-  };
-
-  return [analyticsData, mock];
-};
-
 const appName = 'Codebridge Unit Test';
-const EXPECTED_NEXT_FOLDER_ID = '4';
+const EXPECTED_NEXT_FOLDER_ID = '6';
 
 describe('openNewFolderPrompt', function () {
-  it('can validateNewFolderName', function () {
-    expect(
-      validateNewFolderName({
-        folderName: '',
-        parentId: DEFAULT_FOLDER_ID,
-        projectFolders: testProject.folders,
-      })
-    ).toEqual(undefined);
-
-    expect(
-      validateNewFolderName({
-        folderName: '@',
-        parentId: DEFAULT_FOLDER_ID,
-        projectFolders: testProject.folders,
-      })
-    ).toEqual(codebridgeI18n.invalidNameError());
-
-    expect(
-      validateNewFolderName({
-        folderName: 'testfolder1',
-        parentId: DEFAULT_FOLDER_ID,
-        projectFolders: testProject.folders,
-      })
-    ).toEqual(codebridgeI18n.folderExistsError());
-  });
   it('can successfully add a new folder to root', async function () {
     const [analyticsData, sendCodebridgeAnalyticsEvent] = getAnalyticsMock();
     const newFolderName = 'valid_folder_name';
@@ -89,6 +34,27 @@ describe('openNewFolderPrompt', function () {
 
     await openNewFolderPrompt({
       parentId,
+      appName,
+      dialogControl: getDialogControlMock(newFolderName),
+      newFolder: newFolderDataMock,
+      projectFolders: testProject.folders,
+      sendCodebridgeAnalyticsEvent,
+    });
+
+    expect(newFolderData.parentId).toEqual(parentId);
+    expect(newFolderData.name).toEqual(newFolderName);
+    expect(newFolderData.id).toEqual(EXPECTED_NEXT_FOLDER_ID);
+    expect(analyticsData.event).toEqual(EVENTS.CODEBRIDGE_NEW_FOLDER);
+  });
+
+  it('can successfully implicitly add a new folder to root', async function () {
+    const [analyticsData, sendCodebridgeAnalyticsEvent] = getAnalyticsMock();
+    const newFolderName = 'valid_folder_name';
+    const parentId = DEFAULT_FOLDER_ID;
+
+    const [newFolderData, newFolderDataMock] = getNewFolderMock(parentId);
+
+    await openNewFolderPrompt({
       appName,
       dialogControl: getDialogControlMock(newFolderName),
       newFolder: newFolderDataMock,