+++ apps/src/componentLibrary/slider/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.1](https://github.com/code-dot-org/code-dot-org/pull/61724)
+
+* added hideValue prop
+* hide label section if no value nor label displayed
+
+
+## [0.2.0](https://github.com/code-dot-org/code-dot-org/pull/61560)
+
+* implemented component
+* added tests
+* added storybook
+* component is now ```Ready for Dev```
+
+
 ## [0.1.0](https://github.com/code-dot-org/code-dot-org/pull/61143)
 
 * created component's skeleton